Fix [Bug] Splitting Dataset when shards < n hangs or errors (#22559)
